+++ xla/pjrt/c/pjrt_c_api_wrapper_impl.h
@@ -172,6 +172,10 @@ struct PJRT_Executable {
   std::vector<PJRT_Layouts_MemoryLayout> out_layouts;
   std::vector<PJRT_Layouts_MemoryLayout*> out_layouts_pointers;
 
+  bool parameter_layouts_ran ABSL_GUARDED_BY(mutex) = false;
+  std::vector<PJRT_Layouts_MemoryLayout> parameter_layouts;
+  std::vector<PJRT_Layouts_MemoryLayout*> parameter_layouts_pointers;
+
   bool memory_kind_ran ABSL_GUARDED_BY(mutex) = false;
   std::vector<const char*> memory_kinds;
   std::vector<size_t> memory_kind_sizes;